+++ springboot/src/main/kotlin/com/goliath/emojihub/springboot/controller/UserController.kt
@@ -1,17 +1,15 @@
 package com.goliath.emojihub.springboot.controller
 
-import com.goliath.emojihub.springboot.model.User
+import com.goliath.emojihub.springboot.dto.UserDto
 import com.goliath.emojihub.springboot.service.UserService
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 
 @RestController
-@RequestMapping("/")
+@RequestMapping("/api/user")
 class UserController (private val userService: UserService) {
-    @GetMapping("/users")
-    fun getUsers(): ResponseEntity<List<User>> {
+    @GetMapping
+    fun getUsers(): ResponseEntity<List<UserDto>> {
         return ResponseEntity.ok(userService.getUsers())
     }
 }
