+++ springboot/src/main/kotlin/com/goliath/emojihub/springboot/service/UserService.kt
@@ -1,7 +1,16 @@
 package com.goliath.emojihub.springboot.service
 
-import com.goliath.emojihub.springboot.model.User
+import com.goliath.emojihub.springboot.dao.UserDao
+import com.goliath.emojihub.springboot.dto.UserDto
+import org.springframework.stereotype.Service
 
 interface UserService {
-    fun getUsers(): List<User>
+    fun getUsers(): List<UserDto>
+}
+
+@Service
+class UserServiceImpl(private val userDao: UserDao) : UserService {
+    override fun getUsers(): List<UserDto> {
+        return userDao.getUsers()
+    }
 }
